pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "judo-framework"
version = "1.2.8"
description = "A comprehensive API testing framework for Python, inspired by Karate Framework"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Felipe Farias - CENTYC", email = "felipe.farias@centyc.cl"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Testing",
    "Topic :: Internet :: WWW/HTTP",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.28.0",
    "jsonpath-ng>=1.5.3",
    "pyyaml>=6.0",
    "jsonschema>=4.17.0",
    "faker>=18.0.0",
    "pytest>=7.0.0",
    "behave>=1.2.6",
    "jinja2>=3.1.0",
    "websocket-client>=1.5.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
crypto = [
    "cryptography>=40.0.0",
]
xml = [
    "lxml>=4.9.0",
]
browser = [
    "selenium>=4.8.0",
    "playwright>=1.32.0",
]
full = [
    "cryptography>=40.0.0",
    "lxml>=4.9.0",
    "selenium>=4.8.0",
    "playwright>=1.32.0",
]

[project.scripts]
judo = "judo.cli:main"

[project.urls]
Homepage = "https://github.com/judo-framework/judo"
Documentation = "https://judo-framework.readthedocs.io"
Repository = "https://github.com/judo-framework/judo"
Issues = "https://github.com/judo-framework/judo/issues"
"CENTYC" = "https://www.centyc.cl"
"Author" = "https://www.centyc.cl"

[tool.setuptools.packages.find]
where = ["."]
include = ["judo*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]